Filtro y verificacion al eliminar

+++ Controllers/RoomsViewModel.cs
@@ -5,6 +5,7 @@ using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Runtime.Intrinsics.Arm;
 using System.Windows;
+using System.Windows.Controls;
 using System.Windows.Input;
 using System.Xml.Linq;
 
@@ -29,9 +30,55 @@ public class RoomsViewModel : BaseViewModel
     // La lista de todas las rooms
     public ObservableCollection<RoomModel> Rooms { get; } = new();
 
+    // Filtos
+    //Nombre
+    private string _searchName;
+    public string SearchName
+    {
+        get => _searchName;
+        set
+        {
+            _searchName = value;
+            OnPropertyChanged(nameof(SearchName));
+            _ = LoadRooms();
+        }
+    }
+
+    //Ocupacion
+    private string _filterOccuped;
+    public string FilterOccuped
+    {
+        get => _filterOccuped;
+        set
+        {
+            _filterOccuped = value;
+            OnPropertyChanged(nameof(FilterOccuped));
+            _ = LoadRooms();
+        }
+    }
+
+    public List<string> OccupedFilters { get; } = new()
+    {
+        "", "Libre", "Ocupado"
+    };
+    private bool? ParseOccupedFilter()
+    {
+        return FilterOccuped switch
+        {
+            "Libre" => false,
+            "Ocupado" => true,
+            _ => null
+        };
+    }
+
+
+    //Cargar las habitaciones
     public async Task LoadRooms()
     {
-        var result = await _roomService.GetAllRooms();
+        bool? occuped = ParseOccupedFilter();
+        string name = SearchName;
+
+        var result = await _roomService.GetAllRooms(name, occuped);
 
         if (result.Success && result.Data != null)
         {
@@ -136,16 +183,22 @@ public class RoomsViewModel : BaseViewModel
             if (MessageBox.Show("¿Seguro?", "Confirmación", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
             {
                 //ELIMINAR (DELETE)
-                var result = await _roomService.DeleteRoom(CurrentRoom.RoomId);
-
-                if (!result.Success)
+                if (!CurrentRoom.Occuped)
                 {
-                    MessageBox.Show(result.Error?.Message ?? "Error eliminando habitación");
+                    var result = await _roomService.DeleteRoom(CurrentRoom.RoomId);
+                    if (!result.Success)
+                    {
+                        MessageBox.Show(result.Error?.Message ?? "Error eliminando habitación");
+                        return;
+                    }
+
+                    await LoadRooms();
+                    NavigationViewModel.Instance.BackCommand.Execute(null);
+                }else{
+                    MessageBox.Show("No se puede eliminar habitaciones reservadas", "Habitacion Ocupada", MessageBoxButton.OK, MessageBoxImage.Error);
                     return;
                 }
-
-                await LoadRooms();
-                NavigationViewModel.Instance.BackCommand.Execute(null);
+                
             }
             else { return; }   
         }
@@ -161,5 +214,7 @@ public class RoomsViewModel : BaseViewModel
     private RelayCommand deleteCommand;
     public ICommand DeleteCommand => deleteCommand ??= new RelayCommand(Delete);
 
+
+  
 
 }
